+++ infra/init/10-enable-hypopg.sql
@@ -1,4 +1,1 @@
--- Enable HypoPG extension if available
 CREATE EXTENSION IF NOT EXISTS hypopg;
-
-